Extract Moose::Kokkos::DofSpace from Moose::Kokkos::System
A Kokkos vector is indexed by a system's DOF layout: how many DOFs the process
owns and ghosts, the map from an element's local DOF indices to system-local and
global indices, and the DOF indices exchanged with each other process. That
layout is now a class of its own, which Moose::Kokkos::System inherits and a
p-multigrid level space can own on its own, since a level carries a function
space and none of the tagged vectors, matrices or residual objects a system does.

The layout holds the data rather than declaring an interface: the DOF-count and
element-DOF-index accessors are KOKKOS_FUNCTION and system objects are copied to
device, where a virtual override of a host-only virtual is not usable.
