+++ framework/include/kokkos/systems/KokkosSystem.h
@@ -12,12 +12,9 @@
 #include "KokkosTypes.h"
 #include "KokkosVector.h"
 #include "KokkosMatrix.h"
+#include "KokkosDofSpace.h"
 #include "PerfGraphInterface.h"
-#include "KokkosMesh.h"
 
-#include "libmesh/communicator.h"
-
-class MooseMesh;
 class SystemBase;
 
 namespace Moose::Kokkos
@@ -25,9 +22,10 @@ namespace Moose::Kokkos
 
 /**
  * The Kokkos base system class. Each system in MOOSE has a corresponding Kokkos
- * base system.
+ * base system. On top of the system's DOF layout it carries the tagged vectors and matrices the
+ * residual objects assemble into.
  */
-class System : public PerfGraphInterface, public MeshHolder
+class System : public PerfGraphInterface, public DofSpace
 {
 public:
   /**
@@ -131,30 +129,6 @@ public:
   ///@}
 
   /**
-   * Get the libMesh DOF map
-   * @returns The libMesh DOF map
-   */
-  const libMesh::DofMap & getDofMap() const { return _dof_map; }
-
-  /**
-   * Get the libMesh communicator
-   * @returns The libMesh communicator
-   */
-  const Parallel::Communicator & getComm() const { return _comm; }
-
-  /**
-   * Get the list of local DOF indices to communicate
-   * @returns The list of local DOF indices to communicate
-   */
-  const Array<Array<dof_id_type>> & getLocalCommList() const { return _local_comm_list; }
-
-  /**
-   * Get the list of ghost DOF indices to communicate
-   * @returns The list of ghost DOF indices to communicate
-   */
-  const Array<Array<dof_id_type>> & getGhostCommList() const { return _ghost_comm_list; }
-
-  /**
    * Get the sparisty pattern data
    * @returns The sparisty pattern data
    */
@@ -184,56 +158,6 @@ public:
    * @returns Whether the matrix tag is active
    */
   KOKKOS_FUNCTION bool isMatrixTagActive(TagID tag) const { return _matrix_tag_active[tag]; }
-
-  /**
-   * Get the number of local DOFs
-   * @returns The number of local DOFs
-   */
-  KOKKOS_FUNCTION dof_id_type getNumLocalDofs() const { return _num_local_dofs; }
-
-  /**
-   * Get the number of ghost DOFs
-   * @returns The number of ghost DOFs
-   */
-  KOKKOS_FUNCTION dof_id_type getNumGhostDofs() const { return _num_ghost_dofs; }
-
-  /**
-   * Get the local DOF index of a variable for an element
-   * @param elem The contiguous element ID
-   * @param i The element-local DOF index
-   * @param var The variable number
-   * @returns The local DOF index
-   */
-  KOKKOS_FUNCTION dof_id_type getElemLocalDofIndex(ContiguousElementID elem,
-                                                   unsigned int i,
-                                                   unsigned int var) const
-  {
-    return _local_elem_dof_index[var](i, elem);
-  }
-
-  /**
-   * Get the global DOF index of a variable for an element
-   * @param elem The contiguous element ID
-   * @param i The element-local DOF index
-   * @param var The variable number
-   * @returns The global DOF index
-   */
-  KOKKOS_FUNCTION dof_id_type getElemGlobalDofIndex(ContiguousElementID elem,
-                                                    unsigned int i,
-                                                    unsigned int var) const
-  {
-    return _local_to_global_dof_index[_local_elem_dof_index[var](i, elem)];
-  }
-
-  /**
-   * Get the global DOF index of a local DOF index
-   * @param dof The local DOF index
-   * @returns The global DOF index
-   */
-  KOKKOS_FUNCTION dof_id_type localToGlobalDofIndex(dof_id_type dof) const
-  {
-    return _local_to_global_dof_index[dof];
-  }
 
   /**
    * Get a tagged Kokkos vector
@@ -281,57 +205,12 @@ protected:
   SystemBase & _system;
 
   /**
-   * Reference of the MOOSE mesh
-   */
-  const MooseMesh & _mesh;
-
-  /**
-   * Reference of the libMesh DOF map
-   */
-  const libMesh::DofMap & _dof_map;
-
-  /**
-   * Reference of the libMesh communicator
-   */
-  const Parallel::Communicator & _comm;
-
-  /**
-   * Number of variables
-   */
-  const unsigned int _num_vars;
-
-  /**
-   * Number of local DOFs
-   */
-  const dof_id_type _num_local_dofs;
-
-  /**
-   * Number of ghost DOFs
-   */
-  const dof_id_type _num_ghost_dofs;
-
-  /**
    * Kokkos vectors and matrices on device
    */
   ///@{
   Array<Vector> _vectors;
   Array<Matrix> _matrices;
   ///@}
-
-  /**
-   * Local element DOF indices of each variable
-   */
-  Array<Array2D<dof_id_type>> _local_elem_dof_index;
-
-  /**
-   * Map from local DOF index to global DOF index
-   */
-  Array<dof_id_type> _local_to_global_dof_index;
-
-  /**
-   * Maximum number of DOFs per element for each variable
-   */
-  Array<unsigned int> _max_dofs_per_elem;
 
   /**
    * Whether each variable is active on subdomains
@@ -360,14 +239,6 @@ protected:
   Array<bool> _matrix_tag_active;
   ///@}
 
-  /**
-   * List of DOFs to send and receive
-   */
-  ///@{
-  Array<Array<dof_id_type>> _local_comm_list;
-  Array<Array<dof_id_type>> _ghost_comm_list;
-  ///@}
-
 private:
   /**
    * Setup variable data
@@ -375,9 +246,9 @@ private:
   void setupVariables();
 
   /**
-   * Setup DOF data
+   * Allocate the tagged vector, matrix and tag activity data
    */
-  void setupDofs();
+  void setupTags();
 
   /**
    * Setup sparsity data
